missed a couple of those std::vecor<foo> vs. foo[] cleanups

+++ wbcnet/wbcnet/TaskAtomizer.hpp
@@ -59,7 +59,7 @@ namespace wbcnet {
   */
   class TaskAtomizer
   {
-    typedef unsigned long long bitmap_t;
+    typedef unsigned long bitmap_t;
     
   public:
     typedef enum {
